requirements.txt
numpy==1.24.3
pandas==1.5.3               
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.2.2
scipy==1.10.1
nltk==3.8.1
tqdm==4.65.0
torch==2.0.1
transformers==4.30.2          
sentencepiece==0.1.99
sacremoses==0.0.53
tensorboard==2.13.0
protobuf==3.20.0              
streamlit==1.27.1             
streamlit-autorefresh==1.0.1
python-dotenv==1.0.0          
requests==2.31.0
